Update palindrome.py
I thought it would be good if value entered by the user has all its character in the same case.  Also, if the result would be much more impressive if you give the proper information(i.e. the value entered by the user is palindrome or is not). Hope you would consider it.